PRDT-190 adding facility details page

+++ src/app/services/facility.service.ts
@@ -15,22 +15,26 @@ export class FacilityService {
     private dataService: DataService,
     private loggerService: LoggerService,
     private apiService: ApiService,
-    private loadingService: LoadingService
+    private loadingService: LoadingService,
   ) { }
 
-  async getFacility(fcCollectionId: string, facilityType?: string, facilityId?: string) {
-    const queryParams = null;
+  async getFacility(fcCollectionId: string, facilityType?: string, facilityId?: string, getActivities = false) {
+    const queryParams = {};
     if (facilityType) {
       queryParams['facilityType'] = facilityType;
     }
     if (facilityId) {
       queryParams['facilityId'] = facilityId;
     }
+    if (getActivities) {
+      queryParams['fetchActivities'] = true;
+    }
 
     try {
       this.loadingService.addToFetchList(Constants.dataIds.FACILITY_DETAILS_RESULT);
       const res = (await lastValueFrom(this.apiService.get(`facilities/${fcCollectionId}`, queryParams)))['data'];
       this.dataService.setItemValue(Constants.dataIds.FACILITY_DETAILS_RESULT, res);
+      if (getActivities)
       this.loadingService.removeFromFetchList(Constants.dataIds.FACILITY_DETAILS_RESULT);
     } catch (error) {
       this.loggerService.error(error);
